Version 1.0.7 (#92)
* Rework integration tests and run tests on runner (#59)

* Delete current IT files
* Added test flows
* Added ScargoRunner usage
* Run tests on runner without docker
* Checkout submodules for tests

---------


* Implement passing of CLI options to scargo docker command.

Add unit tests for --rm option.

* Run file formatter.

* Pass docker options as list instead of string to scargo_docker function.

* Fix pylint errors in sc_docker.py

* Fix parameter order when invoking 'docker-compose run' and 'docker exec'

* Update usage documentation of scargo docker command.

* In sc_docker.py subprocess run commands use check=True flag.

* Run scargo update

* Do not run docker inside docker (#60)

* Update github templates (#54)

* Update issue templates

* Update feature_request.md

* Update bug_report.md

* Feature/extra cmake flags in tests (#61)

* Implement --command option for scargo docker run command. (#62)

Add unit test.
Add info to scargo-docker.rst.

* Scargo update updates only main test cmake (#63)

* Add missing comma in readme

* Fix stm32 debug (#64)

* Add configuration option to exclude paths from doc generation

* Eliminate dynamic shell commands

* format

* fix for env variable

* format

* fix

* use typer.launch instead of platform-specific commands

* Release 1.0.5 (#68)

* Release 1.0.5
* Use mocks in docker unit tests
* Use local scargo for integration tests

* fix typing errors

* adjust to strict mypy

* add annotations in tests

* add annotations in common_dev

* add annotations in run.py and clean.py

* add mypy to project

* run format

* update requirements

* update docker image version

* fix critical bug in docker_gen (#76)

* fix

* fix unit test

* run format

* add pre-commit and run pre-commit-hooks linting and checks

add pre-commit and run pre-commit-hooks linting and checks

* add pyupgrade to pre-commit and run it

* add black to pre-commit and run it

* add isort to pre-commit and run it

* add flake8 to pre-commit and run it

* fix scargo gen

* move fixture to conftest

* address review comments

* remove language_version setting for black

* add typing in Jinja2 templates

* add config validation for stm32/esp32

* run format

* add missing types after merge

* add mypy to pre-commit

* refactor mock_gen

* refactor check functions in sc_src

* fix

* use more pathlib instead of os.path

* fix

* raise pylint score to 9.9

* update threshold in CI

* move cli from main __init__

* fix cli imports

* Use EXCLUDE_PATTERNS for doxygen exclusions (#83)

* add exception instead of ignoring a useful pylint rule

* move and rename modules

* split sc_src

* Add --port option to scargo flash (#86)

* Add --port option to scargo flash

I've verified that both parttool.py and esptool.py have the --port option, however st-flash does not have anything like this, so it seems that it can only be supported for esp32 targets.

* move command update to cli.py

* remove leftover imports

* Resolve review comments.

---------


* Tests/docker pypi integration test (#88)

* Add it new/update for pypi scargo version in docker

* Version 1.0.7

---------

+++ scargo/commands/clean.py
@@ -7,8 +7,8 @@ import shutil
 from pathlib import Path
 from typing import Optional
 
-from scargo.scargo_src.sc_logger import get_logger
-from scargo.scargo_src.utils import get_project_root
+from scargo.logger import get_logger
+from scargo.path_utils import get_project_root
 
 
 def _case_insensitive_find_dir(source_dir: Path, dirname: str) -> Optional[Path]:
@@ -28,7 +28,8 @@ def scargo_clean() -> None:
     source_directories = [project_path, test_dir]
 
     for source_dir in source_directories:
-        build_dir = _case_insensitive_find_dir(source_dir, "build")
-        if build_dir and build_dir.exists():
-            shutil.rmtree(build_dir)
-            logger.info("Removed %s", build_dir)
+        if source_dir:
+            build_dir = _case_insensitive_find_dir(source_dir, "build")
+            if build_dir and build_dir.exists():
+                shutil.rmtree(build_dir)
+                logger.info("Removed %s", build_dir)
